+++ src/main/java/com/mmsdivino/user/User.java
@@ -1,19 +1,19 @@
 package com.mmsdivino.user;
 
 public class User {
-    private Integer id;
+    private long id;
     private Double rating;
 
-    public User(Integer id, Double rating) {
+    public User(long id, Double rating) {
         this.id = id;
         this.rating = rating;
     }
 
-    public Integer getId() {
+    public long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(long id) {
         this.id = id;
     }
